+++ src/components/screens/index.js
@@ -3,4 +3,5 @@ export * from './EventRecord';
 export * from './InfoCard';
 export * from './PointOfInterest';
 export * from './PriceCard';
-export * from './TimeCard';
+export * from './OpeningTimesCard';
+export * from './OperatingCompanyInfo';