+++ src/config/styles/index.js
@@ -1,5 +1,7 @@
+import { baseFontStyle } from './baseFontStyle';
 import { html } from './html';
 
 export const styles = {
+  baseFontStyle,
   html
 };