+++ src/config/texts.js
@@ -19,7 +19,8 @@ export const texts = {
     location: 'Anfahrt',
     openingTime: 'Öffnungszeiten',
     prices: 'Preise',
-    description: 'Beschreibung'
+    description: 'Beschreibung',
+    operatingCompany: 'Anbieter'
   },
   screenTitles: {
     home: appJson.expo.name